+++ CHANGELOG.md
@@ -7,6 +7,15 @@ First release.
 > This section will be removed after the beta phase. <br>
 > Note that semantic versioning rules are not strictly followed during this phase.
 
+- v0.2.1: `load()`, `source` and `lazyLoad` now support an extended sytax :
+  ```js
+  source: {
+    url: "path/to/request",
+    params: {},  // key/value pairs converted to URL parameters
+    body: {},    // key/value pairs converted to JSON body (defaults to method POST)
+    options: {}, // passed to `fetch(url, OPTIONS)`
+  }
+  ```
 - v0.2.1: Add  `node.setIcon()`.
 - v0.2.1: Add  `tree.` and `node.sortChildren()`.
 - v0.2.1: Removed `tree.updateColumns()`. Use `tree.setModified(ChangeType.colStructure)` instead.